+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Changes
 -------
 
 * Added the AlertList Panel support in grafanalib/core
+* Added the support for mixed data sources  
 
 
 0.5.2 (unreleased)