+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
 Changes
 -------
 
+* Fixed a bug that was losing any legends on a PromGraph
 * Added the AlertList Panel support in grafanalib/core
 * Added the support for mixed data sources