+++ CHANGELOG.rst
@@ -9,7 +9,8 @@ Changes
 
 * Fixed a bug that was losing any legends on a PromGraph
 * Added the AlertList Panel support in grafanalib/core
-* Added the support for mixed data sources  
+* Added the support for mixed data sources
+* Template now supports 'type' attribute
 
 
 0.5.2 (unreleased)